LPD-74957 LPD-92305 portal-search-elasticsearch8-impl Extract InferenceEndpointInfoFetcher to share the GET _inference fetch https://liferay.atlassian.net/browse/LPD-74957
